+++ packages/editor/src/utils/with-change-detection/test/index.js
@@ -65,6 +65,20 @@ describe( 'withChangeDetection()', () => {
 		expect( state ).toEqual( { count: 1, isDirty: false } );
 	} );
 
+	it( 'should allow ignore function predicate as option', () => {
+		const reducer = withChangeDetection( {
+			isIgnored: ( action ) => action.type === 'INCREMENT',
+		} )( originalReducer );
+
+		let state;
+
+		state = reducer( undefined, {} );
+		expect( state ).toEqual( { count: 0, isDirty: false } );
+
+		state = reducer( deepFreeze( state ), { type: 'INCREMENT' } );
+		expect( state ).toEqual( { count: 1, isDirty: false } );
+	} );
+
 	it( 'should preserve isDirty into non-resetting non-reference-changing types', () => {
 		const reducer = withChangeDetection( { resetTypes: [ 'RESET' ] } )( originalReducer );